ai-engine/requirements.txt
fastapi==0.115.2
python-multipart==0.0.9
pypdf==5.0.1
groq==0.11.0
python-dotenv==1.0.1
